Switched for urxvt to termite.
It turns out that sink index's in 'pacmd list-sinks', are not necessarily in increments of 1. So headphones are index 1, and HDMI is index 4. Go figure.

Changed the for-loop as to not assume that sink_no can be incremented, and instead gmatched all index's and iterate through them, alongside the volumue increase value.